Refactored worker.js code for greater concurrency
